www/url_handlers: Fix a couple of compiler warnings
When compiling with GCC 9.2.1 under Fedora 31 I got the following
compiler warnings

  CC	url_handlers.o
url_handlers.c: In function ‘admin_pending_activations’:
url_handlers.c:761:33: warning: ‘%lu’ directive output may be truncated writing between 1 and 20 bytes into a region of size 3 [-Wformat-truncation=]
  761 |   snprintf(item, sizeof(item), "%lu", i);
      |                                 ^~~
url_handlers.c:761:32: note: directive argument in the range [0, 18446744073709551614]
  761 |   snprintf(item, sizeof(item), "%lu", i);
      |                                ^~~~~
In file included from /usr/include/stdio.h:867,
                 from url_handlers.c:15:
/usr/include/bits/stdio2.h:67:10: note: ‘__builtin___snprintf_chk’ output between 2 and 21 bytes into a destination of size 3
   67 |   return __builtin___snprintf_chk (__s, __n, __USE_FORTIFY_LEVEL - 1,
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   68 |        __bos (__s), __fmt, __va_arg_pack ());
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

and

url_handlers.c: In function ‘approve_receipts’:
url_handlers.c:1533:33: warning: ‘%lu’ directive output may be truncated writing between 1 and 20 bytes into a region of size 3 [-Wformat-truncation=]
 1533 |   snprintf(item, sizeof(item), "%lu", i);
      |                                 ^~~
url_handlers.c:1533:32: note: directive argument in the range [0, 18446744073709551614]
 1533 |   snprintf(item, sizeof(item), "%lu", i);
      |                                ^~~~~
In file included from /usr/include/stdio.h:867,
                 from url_handlers.c:15:
/usr/include/bits/stdio2.h:67:10: note: ‘__builtin___snprintf_chk’ output between 2 and 21 bytes into a destination of size 3
   67 |   return __builtin___snprintf_chk (__s, __n, __USE_FORTIFY_LEVEL - 1,
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   68 |        __bos (__s), __fmt, __va_arg_pack ());
      |        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

These stem from converting a row number (returned from MySQL) which is
an unsigned long into a char[3].

These rows are limited in the MySQL query and will be under 100 so will
fit into the char array including the nul byte.

To fix this, use '%hhu' as the snprintf(3) format specifier and cast the
row number to a uint8_t.

+++ www/url_handlers.c
@@ -4,7 +4,7 @@
  * Copyright (C) 2011-2013	OpenTech Labs
  *				Andrew Clayton <andrew@digital-domain.net>
  *
- *		 2016		Andrew Clayton <andrew@digital-domain.net>
+ *		 2016, 2020	Andrew Clayton <andrew@digital-domain.net>
  *
  * Released under the GNU Affero General Public License version 3.
  * See COPYING
@@ -758,7 +758,7 @@ static void admin_pending_activations(void)
 
 		lf_set_var(f, "uid", get_var(db_row, "uid"), NULL);
 		lf_set_var(f, "akey", get_var(db_row, "akey"), NULL);
-		snprintf(item, sizeof(item), "%lu", i);
+		snprintf(item, sizeof(item), "%hhu", (uint8_t)i);
 		lf_set_var(f, "item", item, NULL);
 
 		do_zebra(f, i, "listzebra");
@@ -1530,7 +1530,7 @@ static void approve_receipts(void)
 		lf_set_var(f, "reason", get_var(db_row, "reason"), de_xss);
 		lf_set_var(f, "id", get_var(db_row, "id"), NULL);
 
-		snprintf(item, sizeof(item), "%lu", i);
+		snprintf(item, sizeof(item), "%hhu", (uint8_t)i);
 		lf_set_var(f, "item", item, NULL);
 
 		lf_set_row(f, "table");
